Bake sets
Implemented bake sets and refactored the code
