--- _config.yml
@@ -1,5 +1,0 @@
-remote_theme: pages-themes/slate@v0.2.0
-plugins:
-- jekyll-remote-theme # add this line to the plugins list if you already have one
-markdown: kramdown
-title: Substance bake prep tool for Blender
