--- assets/css/style.scss
@@ -1,58 +1,0 @@
----
----
-
-@import "{{ site.theme }}";
-nav {
-  width: 230px;
-  position: fixed;
-  top: 220px;
-  left:50%;
-  margin-left:-580px;
-  text-align: right;
-
-  ul {
-    list-style: none;
-    list-style-image:none;
-    font-size: 14px;
-    line-height:24px;
-
-    li {
-      padding: 5px 0px;
-      line-height: 16px;
-      // padding-right:17px;
-      // position:relative;
-      // right:-12px;
-
-      &.tag-h1 {
-        font-size: 1.2em;
-
-        a {
-          font-weight: bold;
-          color: #333;
-        }
-
-        + .tag-h2 {
-
-        }
-      }
-
-      &.tag-h2 {
-
-        + .tag-h1 {
-          margin-top:10px;
-        }
-      }
-    }
-
-    a {
-      color: #666;
-
-      &:hover { color: #999; }
-    }
-
-    // .active {
-    //   border-right:solid 4px #39C;
-    //   padding-right:13px;
-    // }
-  }
-}
